opnsense/bsdinstaller: select disk for quick/easy install
Allows us to zap the confirmation dialog and install to a different
media in case there is more than one disk or flash card available.
Will need more bsdlabel trickery to really fly, but we're getting
there...  Ride port revision increases for now, but will be wrapped
up as 2.2 eventually.