opnsense/bsdinstaller: switch to async mount for speed (#5)
Ad pointed out that SD cards can take a very long time to write
due to FreeBSD handling disks `noasync' by default.  While it is
labeled `dangerous', it is fine for the bsdinstaller's use case
as we bootstrap a system and are not overly concerned with data
loss due to sudden power outages.  The final system will still
be `noasync' so the world is still a safe place.