opnsense/bsdinstaller: transform post-install tasks into a script
The previous would read one line at a time executing those, creating
a lot of fuzz while trying to work on post-install customisation.
Eventually, post-install should be embedded into LUA, but for now
we have two callers and cleargpt.sh is there too so at least be more
consistent for the time being.