opnsense/bsdinstaller: bring back swap activate; closes #17
This was removed in commit d6fcf59ae.  Unfortunately, the base
system is incapable of cloning a file or directory that is greater
than the amount of available system memory.  For embedded devices
or sparse VM configurations this will fail as the installer needs
some 700ish MB in order to install fine.

Previously, this wasn't considered a viable approach as this will
still fail on embedded installations where no swap is configured.

For this we have nano-style images, these don't require an install.