Framework: compat with old SSP_CLFAGS behaviour
Keep behaviour the way it was.  misc/mc-light configure fails with
-fstack-protector-all, but not without SSP, -fstack-protector
and -fstack-protector-strong.

We should move to -fstack-protector-strong if only for follwing
FreeBSD ports, but since the impact on the ports is hard to
follow in OPNsense and would creep into package updates of stable
packages in 19.1.x it's better to wait for the clean reinstall of
19.7 packages during the upgrade.

CC: @lattera