opnsenses/filterlog: avoid extracting IPv6 options that are never parsed
Basically this deletes a lot of esoteric code that was copied from
tcpdump when filterlog was created.  The options format is confusing
and not supported by any parser we know of.

Instead just jump through all options to see if we have something
that looks like the right protocol (TCP, UDP or CARP/VRRP).

PR: https://github.com/opnsense/core/issues/5016