www/lighttpd: partially sync with upstream
Update, but keep the old pcre library for a bit longer since the
pcre2 implementation assumes JIT is compiled into the libary when
in reality JIT support is turned off (due to HardenedBSD).

Eventually we will gain JIT in 22.1 for to switch over to pcre2,
but in the meantime to not break anything obvious this route here
seems sensible.

Taken from: FreeBSD