www/phalcon: update to 5.0.0beta2
The software has come a long way.  Use the phalcon4 port as
a base but stop restricting architectures since this version
offers a unified build directory and a lot of architectures
have been added in the past which seem to work well.

Version 3 does not support a higher version than PHP 7.3 and
that is already out of support so update the port with the
incompatible version and keep phalcon4 out of this.  Being
a beta already this should be good to use and everyone else
should have really migrated to phalcon4 already.