sysutils/syslog-ng: LibreSSL again
It's dead, Jim.  For now fail on use of sigalgs() and client-sigalgs()
which we don't seem to be using at this point.  If we ever do it's the
final straw for LibreSSL.  OpenVPN and Syslog-ng are core components
driving the nails in the coffin for LibreSSL nowadays.  If we make it
to January 2023 that would be great but it could be sooner as OpenSSL
3.0 approaches.