opnsense/ddclient: it was so nice, let's patch it twice
The patch contents of d35d62f was patched for dondomino service,
not duckdns as it was intended.  The reason for this is the
excessive code duplication inside ddclient to the point where
new protocols are copied from existing protocols even though
the code does not change.

We cleverly work around this by feeding the patch twice which
patches the right code on the second try.  You can't make this up!

PR: https://forum.opnsense.org/index.php?topic=32273.0