opnsense/filterlog: rulenr can be -1 as well #177
This dates back to OpenBSD and was never correctly handled in
tcpdump.  There is an internal default rule which may or may
not have the correct action set when e.g. PFRES_SHORT causes
the packet to be dropped before a rule is matched.  Now we leave
the rulenr empty like subrulenr instead of showing "4294967295"
which is probably not an existing rule number.

In the label check simply check the upper and lower bound before
the lookup instead of doing more casts for an implicit compare.