opnsense/update: minor bootstrap improvements and safety belts
We now embed the valid FreeBSD version from the build process and the
build will complain if it can't get any of those values it needs.

OTOH, the plain script that can be downloaded from GitHub will now
simply ignore the FreeBSD version check as that creates issues sometimes.
In a sense it is more dangerous now but fully outside our ecosystem and
immediate reach.