security/ca_root_nss: move openssl/cert.pem to ETCSYMLINK option
If the base system works fine with a direct link also do the
same with ${PREFIX}/openssl/cert.pem instead of prodiving an
overwritable sample file to the effective location of the
security/openssl port directory.

This avoids accidential writes to the effective trust bundle
location when ETCSYMLINK is off.  ${PREFIX}/etc/ssl/cert.pem
remains a better target for a custom bundle location.

Do this at the cost of user-based modifications in the non-sample
location.  Ports have never been hardwired to this location though.

It would be nice if ETCSYMLINK could default to off in a not so
distant future.