opnsense/libxml2-legacy: compat shim for old binary
Provide a shim for the PHP simplexml/dom boot issues that users reported.
This happened because libxml changed SONAME which appears to be customary nowadays
for introducing breakage in the ports tree.

By providing the old binary with the newer libxml2 as a dependency PHP works
pre,mid and post update and even on partial upgrades.