Revert "databases/py-sqlite3: XXX decrease for sync due to Python 3.13 update"
This reverts commit 67873f9df87607639042ee6014eee29edf1bef47.

Python 3.13 isn't ready because py-numpy depends on cython (not cython3) and that
is being deprecated at the moment.  Looks like the main issue is that py-numpy is
at the older major version, which got a new major 2.x release in summer 2024.

We could instead head for 3.12 but that would likely mean more noise than worth
the effort.